pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = []

[tool.poetry.dependencies]
python = "^3.8.11"
python-dotenv = "^0.19.0"
azure-communication-sms = "^1.0.1"
Flask = "1.1.2"
Flask-PyMongo = "2.3.0"
gunicorn = "20.0.4"
requests = "^2.23.0"
sendgrid = "6.3.1"
Flask-RESTful = "0.3.8"
pandas = "1.2.3"
Flask-Cors = "3.0.8"
xkcdpass = "1.17.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
